fixed when running imports

+++ back-end/data-base_functions/deletes/delete_all.ts
@@ -37,5 +37,6 @@ async function deleteAllDocuments() {
 export {deleteAllDocuments};
 
 // Run the function
-deleteAllDocuments();
-
+if(import.meta.main) {
+  deleteAllDocuments();
+}
